pyproject.toml
[tool.poetry]
name = "reach4pitch"
version = "0.0.16"
description = ""
authors = ["ChaikaBogdan", "Aristarhys"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
django-flags = "^5.0.13"
django-simple-captcha = "^0.5.17"
django-rq = "^2.8.1"
gunicorn = "^20.1.0"
psycopg = "^3.1.9"
requests = "^2.31.0"
django-tinymce = "^3.6.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.275"
pre-commit = "^3.3.3"
black = "^23.3.0"
djlint = "^1.31.1"
djhtml = "^3.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "migrations",
]
per-file-ignores = {}

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.11
target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
